// src/app/services/models/offre-emploi-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OffreEmploiDto {
  id?: number;
  competences?: Array<number>;
  description?: string;
  experience?: number;
  localisation?: string;
  recruteurId?: number;
  salaire?: number;
  status?: 'OPEN' | 'CLOSED';
  titre?: string;
}